+++ node-device/node-plug-state.js
@@ -8,8 +8,7 @@ module.exports = function (RED) {
 
     if (device && plug) {
       // Setup a subscriber to the plug's state topic
-      const topic = `servicelocation/${device.uuid}/plug/${plug.uuid}/state`
-      device.subscribe(topic, node)
+      device.subscribe(`plug/${plug.uuid}/state`, node)
 
       // Listen to inputs
       node.on('input', function (msg) {
@@ -19,7 +18,7 @@ module.exports = function (RED) {
 
           // Publish the on or off message
           if (typeof onState !== 'undefined') {
-            device.publish(`servicelocation/${device.uuid}/plug/${plug.uuid}/setstate`, {
+            device.publish(`plug/${plug.uuid}/setstate`, {
               value: onState ? 'ON' : 'OFF',
               since: new Date().getTime(),
               nodeId: plug.uuid